LayoutTests/media/video-dom-autoplay.html

<video controls></video>
<script src=video-test.js></script>
<script>
    testExpected("video.autoplay", false);
    video.autoplay = true;
    testExpected("video.autoplay", true);
    testExpected("video.getAttribute('autoplay')", null, "!=");
    waitForEvent('play', function () { 
        testExpected("video.paused", false);
        endTest();
    } );

    video.src = "content/test.mp4";
</script>
