LayoutTests/media/video-dom-autoplay-expected.txt

EXPECTED (video.autoplay == 'false') OK
EXPECTED (video.autoplay == 'true') OK
EXPECTED (video.getAttribute('autoplay') != 'null') OK
EVENT(play)
EXPECTED (video.paused == 'false') OK
END OF TEST
